+++ domain/schema/package_test.go
@@ -4,11 +4,98 @@
 package schema
 
 import (
+	"context"
+	"database/sql"
+	"fmt"
 	"testing"
+	"time"
 
+	"github.com/juju/collections/set"
+	jc "github.com/juju/testing/checkers"
 	gc "gopkg.in/check.v1"
+
+	"github.com/juju/juju/core/database/schema"
+	databasetesting "github.com/juju/juju/internal/database/testing"
 )
 
 func Test(t *testing.T) {
 	gc.TestingT(t)
+}
+
+type schemaBaseSuite struct {
+	databasetesting.DqliteSuite
+}
+
+// NewCleanDB returns a new sql.DB reference.
+func (s *schemaBaseSuite) NewCleanDB(c *gc.C) *sql.DB {
+	dir := c.MkDir()
+
+	url := fmt.Sprintf("file:%s/db.sqlite3?_foreign_keys=1", dir)
+	c.Logf("Opening sqlite3 db with: %v", url)
+
+	db, err := sql.Open("sqlite3", url)
+	c.Assert(err, jc.ErrorIsNil)
+
+	return db
+}
+
+func (s *schemaBaseSuite) applyDDL(c *gc.C, ddl *schema.Schema) {
+	if s.Verbose {
+		ddl.Hook(func(i int, statement string) error {
+			c.Logf("-- Applying schema change %d\n%s\n", i, statement)
+			return nil
+		})
+	}
+	changeSet, err := ddl.Ensure(context.Background(), s.TxnRunner())
+	c.Assert(err, jc.ErrorIsNil)
+	c.Check(changeSet.Current, gc.Equals, 0)
+	c.Check(changeSet.Post, gc.Equals, ddl.Len())
+}
+
+func (s *schemaBaseSuite) assertExecSQL(c *gc.C, q string, args ...any) {
+	err := s.TxnRunner().StdTxn(context.Background(), func(ctx context.Context, tx *sql.Tx) error {
+		_, err := tx.ExecContext(ctx, q, args...)
+		return err
+	})
+	c.Assert(err, jc.ErrorIsNil)
+}
+
+func (s *schemaBaseSuite) assertExecSQLError(c *gc.C, q string, errMsg string, args ...any) {
+	err := s.TxnRunner().StdTxn(context.Background(), func(ctx context.Context, tx *sql.Tx) error {
+		_, err := tx.ExecContext(ctx, q, args...)
+		return err
+	})
+	c.Assert(err, gc.ErrorMatches, errMsg)
+}
+
+var (
+	internalTableNames = set.NewStrings(
+		"schema",
+		"sqlite_sequence",
+	)
+)
+
+func readEntityNames(c *gc.C, db *sql.DB, entity_type string) []string {
+	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
+	defer cancel()
+
+	tx, err := db.BeginTx(ctx, nil)
+	c.Assert(err, jc.ErrorIsNil)
+
+	rows, err := tx.QueryContext(ctx, `SELECT DISTINCT name FROM sqlite_master WHERE type = ? ORDER BY name ASC;`, entity_type)
+	c.Assert(err, jc.ErrorIsNil)
+	defer func() { _ = rows.Close() }()
+
+	var names []string
+	for rows.Next() {
+		var name string
+		err = rows.Scan(&name)
+		c.Assert(err, jc.ErrorIsNil)
+		names = append(names, name)
+	}
+
+	err = tx.Commit()
+	c.Assert(err, jc.ErrorIsNil)
+
+	return names
 }
